Renames capstone package to univiz
