--- capstone/src/main/java/com/google/capstone/CapstoneBootstrapModule.java
@@ -1,5 +1,0 @@
-package com.google.capstone;
-
-import com.google.inject.AbstractModule;
-
-final class CapstoneBootstrapModule extends AbstractModule {}
