+++ capstone/src/main/java/com/google/univiz/UnivizBootstrapModule.java
@@ -1,0 +1,5 @@
+package com.google.univiz;
+
+import com.google.inject.AbstractModule;
+
+final class UnivizBootstrapModule extends AbstractModule {}
