+++ capstone/src/main/webapp/WEB-INF/web.xml
@@ -11,7 +11,7 @@
     <url-pattern>/*</url-pattern>
   </filter-mapping>
   <listener>
-    <listener-class>com.google.capstone.GuiceServletConfig</listener-class>
+    <listener-class>com.google.univiz.GuiceServletConfig</listener-class>
   </listener>
-  <display-name>Archetype Created Web Application</display-name>
+  <display-name>UniViz Web Application</display-name>
 </web-app>
